requirements.txt
attrs==19.3.0
blinker==1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Click==7.0
coverage==5.1
cryptography==2.9
docutils==0.15.2
dominate==2.4.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
python-dotenv==0.10.4
pytz==2019.3
requests==2.23.0
sentry-sdk==0.14.3
six==1.13.0
ujson==2.0.3
urllib3==1.25.7
visitor==0.1.3
wcwidth==0.1.9
Werkzeug==0.16.0
WTForms==2.2.1
zipp==3.1.0
